Finish create function

+++ ingredient/views.py
@@ -1,27 +1,25 @@
 from django.db import connection
 from django.shortcuts import render, redirect
+import random, string
 
 def create_ingredient(request):
     context = {}
+
     with connection.cursor() as cursor:
         cursor.execute("SET SEARCH_PATH TO SIREST")
-        cursor.execute("""
-            SELECT name
-            FROM INGREDIENT 
-            ORDER BY name DESC
-        """)
-        ingredient = int(cursor.fetchone()[0]) + 1
-        context["name"] = ingredient
-
+        
         if request.method == "POST":
-            name = request.POST.get("name")
-            cursor.execute(f"""
-                    INSERT INTO INGREDIENT VALUES
-                    ('{name}')
-                """)
-            return redirect("/ingredient/")
+            random_id = "".join(('I' + random.choice(string.digits)))
+            ingredient_name = request.POST.get("ingredient_name")
 
-    return render(request, 'c_ingredient.html', context)
+            cursor.execute(f"""
+                INSERT INTO INGREDIENT VALUES
+                ('{random_id}', '{ingredient_name}')
+            """)
+
+            return redirect("/ingredient/")
+        
+    return render(request, "c_ingredient.html", context)
 
 def read_ingredient(request):
     context = {}
@@ -37,11 +35,27 @@ def read_ingredient(request):
 
     return render(request, 'r_ingredient.html', context)
 
+# def delete_ingredient(request):
+#     name = request.session["name"]
+#     with connection.cursor() as cursor:
+#         cursor.execute("SET SEARCH_PATH TO SIREST")
+#         cursor.execute(f"""
+#         DELETE FROM INGREDIENT
+#         WHERE name = '{name}'""")
+#         return redirect(f"/ingredient/")
+
 def delete_ingredient(request):
-    name = request.session["name"]
+    context = {}
+    nama = request.session["nama"]
+
     with connection.cursor() as cursor:
         cursor.execute("SET SEARCH_PATH TO SIREST")
-        cursor.execute(f"""
-        DELETE FROM INGREDIENT
-        WHERE name = '{name}'""")
-        return redirect(f"/ingredient/")
+        
+        if request.method == "POST":
+            cursor.execute(f"""
+                DELETE FROM INGREDIENT
+                WHERE name = '{nama}'
+            """)
+
+            return redirect("/ingredient/")
+    return render(request, "r_ingredient.html", context)
